+++ pkg/resources/database.go
@@ -1,17 +1,14 @@
 package resources
 
 import (
+	"context"
 	"database/sql"
-	"errors"
 	"fmt"
-	"log"
-	"strconv"
-	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/jmoiron/sqlx"
+	"golang.org/x/exp/slices"
 
-	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/snowflake"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
 )
 
 var databaseSchema = map[string]*schema.Schema{
@@ -80,7 +77,6 @@ var databaseSchema = map[string]*schema.Schema{
 			},
 		},
 	},
-	"tag": tagReferenceSchema,
 }
 
 // Database returns a pointer to the resource representing a database.
@@ -98,248 +94,225 @@ func Database() *schema.Resource {
 	}
 }
 
-func createDatabase(d *schema.ResourceData, builder *snowflake.DatabaseBuilder, meta interface{}) error {
-	db := meta.(*sql.DB)
-	q := builder.Create()
-	name := d.Get("name").(string)
-
-	if err := snowflake.Exec(db, q); err != nil {
-		return fmt.Errorf("error creating database %v err = %w", name, err)
-	}
-
-	d.SetId(name)
-
-	return ReadDatabase(d, meta)
-}
-
 // CreateDatabase implements schema.CreateFunc.
 func CreateDatabase(d *schema.ResourceData, meta interface{}) error {
-	// TODO: Migrate database from share and from replica to iterative approach
-	if _, ok := d.GetOk("from_share"); ok {
-		return createDatabaseFromShare(d, meta)
-	}
-
-	if _, ok := d.GetOk("from_replica"); ok {
-		return createDatabaseFromReplica(d, meta)
-	}
-
+	db := meta.(*sql.DB)
+	client := sdk.NewClientFromDB(db)
+	ctx := context.Background()
 	name := d.Get("name").(string)
-	builder := snowflake.NewDatabaseBuilder(name)
+	id := sdk.NewAccountObjectIdentifier(name)
 
-	// Set optionals
+	// Is it a Shared Database?
+	if fromShare, ok := d.GetOk("from_share"); ok {
+		account := fromShare.(map[string]interface{})["provider"].(string)
+		share := fromShare.(map[string]interface{})["share"].(string)
+		shareID := sdk.NewExternalObjectIdentifier(sdk.NewAccountIdentifierFromAccountLocator(account), sdk.NewAccountObjectIdentifier(share))
+		opts := &sdk.CreateSharedDatabaseOptions{}
+		if v, ok := d.GetOk("comment"); ok {
+			opts.Comment = sdk.String(v.(string))
+		}
+		err := client.Databases.CreateShared(ctx, id, shareID, opts)
+		if err != nil {
+			return fmt.Errorf("error creating database %v: %w", name, err)
+		}
+		d.SetId(name)
+		if v, ok := d.GetOk("replication_configuration"); ok {
+			replicationConfiguration := v.([]interface{})[0].(map[string]interface{})
+			accounts := replicationConfiguration["accounts"].([]interface{})
+			accountIDs := make([]sdk.AccountIdentifier, len(accounts))
+			for i, account := range accounts {
+				accountIDs[i] = sdk.NewAccountIdentifierFromAccountLocator(account.(string))
+			}
+			opts := &sdk.AlterDatabaseReplicationOptions{
+				EnableReplication: &sdk.EnableReplication{
+					ToAccounts: accountIDs,
+				},
+			}
+			if ignoreEditionCheck, ok := replicationConfiguration["ignore_edition_check"]; ok {
+				opts.EnableReplication.IgnoreEditionCheck = sdk.Bool(ignoreEditionCheck.(bool))
+			}
+			err := client.Databases.AlterReplication(ctx, id, opts)
+			if err != nil {
+				return fmt.Errorf("error enabling replication for database %v: %w", name, err)
+			}
+		}
+		return ReadDatabase(d, meta)
+	}
+	// Is it a Secondary Database?
+	if primaryName, ok := d.GetOk("from_replica"); ok {
+		primaryID := sdk.NewExternalObjectIdentifierFromFullyQualifiedName(primaryName.(string))
+		opts := &sdk.CreateSecondaryDatabaseOptions{
+			DataRetentionTimeInDays: sdk.Int(d.Get("data_retention_time_in_days").(int)),
+		}
+		err := client.Databases.CreateSecondary(ctx, id, primaryID, opts)
+		if err != nil {
+			return fmt.Errorf("error creating database %v: %w", name, err)
+		}
+		d.SetId(name)
+		// todo: add failover_configuration block
+		return ReadDatabase(d, meta)
+	}
+
+	// Otherwise it is a Standard Database
+	opts := sdk.CreateDatabaseOptions{}
 	if v, ok := d.GetOk("comment"); ok {
-		builder.WithComment(v.(string))
+		opts.Comment = sdk.String(v.(string))
 	}
 
 	if v, ok := d.GetOk("is_transient"); ok && v.(bool) {
-		builder.Transient()
+		opts.Transient = sdk.Bool(v.(bool))
 	}
 
 	if v, ok := d.GetOk("from_database"); ok {
-		builder.Clone(v.(string))
+		opts.Clone = &sdk.Clone{
+			SourceObject: sdk.NewAccountObjectIdentifier(v.(string)),
+		}
 	}
 
 	if v, ok := d.GetOk("data_retention_time_in_days"); ok {
-		builder.WithDataRetentionDays(v.(int))
+		opts.DataRetentionTimeInDays = sdk.Int(v.(int))
 	}
 
-	if v, ok := d.GetOk("tag"); ok {
-		tags := getTags(v)
-		builder.WithTags(tags.toSnowflakeTagValues())
+	err := client.Databases.Create(ctx, id, &opts)
+	if err != nil {
+		return fmt.Errorf("error creating database %v: %w", name, err)
 	}
-
-	// If set, verify parameters are valid and attempt to enable replication
-	if v, ok := d.GetOk("replication_configuration"); ok {
-		replicationConfiguration := v.([]interface{})[0].(map[string]interface{})
-		ignoreEditionCheck := replicationConfiguration["ignore_edition_check"].(bool)
-
-		if !ignoreEditionCheck {
-			return errors.New("error enabling replication - ignore edition check was set to false")
-		}
-		resource := createDatabase(d, builder, meta)
-		if err := enableReplication(d, meta, replicationConfiguration); err != nil {
-			return fmt.Errorf("error enabling replication - account does not exist or System Parameter ENABLE_ACCOUNT_DATABASE_REPLICATION must be set to true, err = %w", err)
-		}
-		return resource
-	}
-
-	return createDatabase(d, builder, meta)
-}
-
-func enableReplication(d *schema.ResourceData, meta interface{}, replicationConfig map[string]interface{}) error {
-	db := meta.(*sql.DB)
-	primaryDBName := d.Get("name").(string)
-	accounts := replicationConfig["accounts"].([]interface{})
-	accountsToEnableReplication := strings.Join(expandStringList(accounts), ", ")
-	enableReplicationStmt := fmt.Sprintf(`ALTER DATABASE "%s" ENABLE REPLICATION TO ACCOUNTS %s`, primaryDBName, accountsToEnableReplication)
-	return snowflake.Exec(db, enableReplicationStmt)
-}
-
-func createDatabaseFromShare(d *schema.ResourceData, meta interface{}) error {
-	in := d.Get("from_share").(map[string]interface{})
-	prov := in["provider"]
-	share := in["share"]
-
-	if prov == nil || share == nil {
-		return fmt.Errorf("from_share must contain the keys provider and share, but it had %+v", in)
-	}
-
-	db := meta.(*sql.DB)
-	name := d.Get("name").(string)
-	builder := snowflake.DatabaseFromShare(name, prov.(string), share.(string))
-
-	if comment, ok := d.GetOk("comment"); ok {
-		builder.WithComment(comment.(string))
-	}
-
-	if err := snowflake.Exec(db, builder.Create()); err != nil {
-		return fmt.Errorf("error creating database %v from share %v.%v err = %w", name, prov, share, err)
-	}
-
 	d.SetId(name)
-
-	return ReadDatabase(d, meta)
-}
-
-func createDatabaseFromReplica(d *schema.ResourceData, meta interface{}) error {
-	sourceDB := d.Get("from_replica").(string)
-
-	db := meta.(*sql.DB)
-	name := d.Get("name").(string)
-	builder := snowflake.DatabaseFromReplica(name, sourceDB)
-
-	if err := snowflake.Exec(db, builder.Create()); err != nil {
-		return fmt.Errorf("error creating a secondary database %v from database %v err = %w", name, sourceDB, err)
-	}
-
-	d.SetId(name)
-
 	return ReadDatabase(d, meta)
 }
 
 func ReadDatabase(d *schema.ResourceData, meta interface{}) error {
 	db := meta.(*sql.DB)
-	dbx := sqlx.NewDb(db, "snowflake")
+	client := sdk.NewClientFromDB(db)
+	ctx := context.Background()
 	name := d.Id()
+	id := sdk.NewAccountObjectIdentifier(name)
 
-	// perform a "show database" command to ensure that the database is actually there.
-	stmt := snowflake.NewDatabaseBuilder(name).Show()
-	row := snowflake.QueryRow(db, stmt)
-	_, err := snowflake.ScanDatabase(row)
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			// If not found, mark resource to be removed from statefile during apply or refresh
-			log.Printf("[DEBUG] database (%s) not found", name)
-			d.SetId("")
-			return nil
-		}
-		return fmt.Errorf("unable to scan row for SHOW DATABASES")
-	}
-
-	// there may be more than one database found, so we need to filter. this could probably be combined with the above query
-	database, err := snowflake.ListDatabase(dbx, name)
+	database, err := client.Databases.ShowByID(ctx, id)
 	if err != nil {
 		return err
 	}
-	if err := d.Set("name", database.DBName.String); err != nil {
+	if err := d.Set("name", database.Name); err != nil {
 		return err
 	}
-	if err := d.Set("comment", database.Comment.String); err != nil {
-		return err
-	}
-
-	i, err := strconv.ParseInt(database.RetentionTime.String, 10, 64)
-	if err != nil {
-		return err
-	}
-	if err := d.Set("data_retention_time_in_days", i); err != nil {
+	if err := d.Set("comment", database.Comment); err != nil {
 		return err
 	}
 
-	// reset the options before reading back from the DB
-	if err := d.Set("is_transient", false); err != nil {
+	if err := d.Set("data_retention_time_in_days", database.RetentionTime); err != nil {
 		return err
 	}
 
-	if opts := database.Options.String; opts != "" {
-		for _, opt := range strings.Split(opts, ", ") {
-			if opt == "TRANSIENT" {
-				if err := d.Set("is_transient", true); err != nil {
-					return err
-				}
-			}
-		}
+	if err := d.Set("is_transient", database.Transient); err != nil {
+		return err
 	}
 
 	return nil
 }
 
 func UpdateDatabase(d *schema.ResourceData, meta interface{}) error {
-	dbName := d.Id()
-	builder := snowflake.NewDatabaseBuilder(dbName)
+	name := d.Id()
+	id := sdk.NewAccountObjectIdentifier(name)
 	db := meta.(*sql.DB)
-
-	// If replication configuration changes, need to update accounts that have permission to replicate database
-	if d.HasChange("replication_configuration") {
-		oldConfig, newConfig := d.GetChange("replication_configuration")
-		newConfigLength := len(newConfig.([]interface{}))
-		oldConfigLength := len(oldConfig.([]interface{}))
-		// Enable replication for any new accounts and disable replication for removed accounts
-		if newConfigLength > 0 {
-			newAccounts := extractInterfaceFromAttribute(newConfig, "accounts")
-			enableQuery := builder.EnableReplicationAccounts(dbName, strings.Join(expandStringList(newAccounts), ", "))
-			if err := snowflake.Exec(db, enableQuery); err != nil {
-				return fmt.Errorf("error enabling replication configuration with statement %v err = %w", enableQuery, err)
-			}
-		}
-
-		if oldConfigLength > 0 {
-			oldAccounts := extractInterfaceFromAttribute(oldConfig, "accounts")
-			var accountsToDisableReplication []interface{}
-			if newConfigLength > 0 {
-				newAccounts := extractInterfaceFromAttribute(newConfig, "accounts")
-				accountsToDisableReplication = builder.GetRemovedAccountsFromReplicationConfiguration(oldAccounts, newAccounts)
-			} else {
-				accountsToDisableReplication = builder.GetRemovedAccountsFromReplicationConfiguration(oldAccounts, nil)
-			}
-			// If accounts were found to be removed, disable replication
-			if len(accountsToDisableReplication) > 0 {
-				disableQuery := builder.DisableReplicationAccounts(dbName, strings.Join(expandStringList(accountsToDisableReplication), ", "))
-				if err := snowflake.Exec(db, disableQuery); err != nil {
-					return fmt.Errorf("error disabling replication configuration with statement %v err = %w", disableQuery, err)
-				}
-			}
-		}
-	}
+	client := sdk.NewClientFromDB(db)
+	ctx := context.Background()
 
 	if d.HasChange("name") {
-		name := d.Get("name")
-		q := builder.Rename(name.(string))
-		if err := snowflake.Exec(db, q); err != nil {
+		newName := d.Get("name").(string)
+		opts := &sdk.AlterDatabaseOptions{
+			NewName: sdk.NewAccountObjectIdentifier(newName),
+		}
+		err := client.Databases.Alter(ctx, id, opts)
+		if err != nil {
 			return fmt.Errorf("error updating database name on %v err = %w", d.Id(), err)
 		}
-		d.SetId(name.(string))
+		d.SetId(newName)
+		id = sdk.NewAccountObjectIdentifier(newName)
 	}
 
 	if d.HasChange("comment") {
-		comment := d.Get("comment")
-		q := builder.ChangeComment(comment.(string))
-		if err := snowflake.Exec(db, q); err != nil {
+		comment := ""
+		if c := d.Get("comment"); c != nil {
+			comment = c.(string)
+		}
+		opts := &sdk.AlterDatabaseOptions{
+			Set: &sdk.DatabaseSet{
+				Comment: sdk.String(comment),
+			},
+		}
+		err := client.Databases.Alter(ctx, id, opts)
+		if err != nil {
 			return fmt.Errorf("error updating database comment on %v err = %w", d.Id(), err)
 		}
 	}
 
 	if d.HasChange("data_retention_time_in_days") {
-		days := d.Get("data_retention_time_in_days")
-		q := builder.ChangeDataRetentionDays(days.(int))
-		if err := snowflake.Exec(db, q); err != nil {
-			return fmt.Errorf("error updating data retention days on %v err = %w", d.Id(), err)
+		days := d.Get("data_retention_time_in_days").(int)
+		opts := &sdk.AlterDatabaseOptions{
+			Set: &sdk.DatabaseSet{
+				DataRetentionTimeInDays: sdk.Int(days),
+			},
+		}
+		err := client.Databases.Alter(ctx, id, opts)
+		if err != nil {
+			return fmt.Errorf("error updating database data retention time on %v err = %w", d.Id(), err)
 		}
 	}
 
-	tagChangeErr := handleTagChanges(db, d, builder)
-	if tagChangeErr != nil {
-		return tagChangeErr
+	// If replication configuration changes, need to update accounts that have permission to replicate database
+	if d.HasChange("replication_configuration") {
+		oldConfig, newConfig := d.GetChange("replication_configuration")
+		newAccounts := newConfig.([]interface{})[0].(map[string]interface{})["accounts"].([]interface{})
+		newAccountIDs := make([]sdk.AccountIdentifier, len(newAccounts))
+		for i, account := range newAccounts {
+			newAccountIDs[i] = sdk.NewAccountIdentifierFromAccountLocator(account.(string))
+		}
+		oldAccounts := oldConfig.([]interface{})[0].(map[string]interface{})["accounts"].([]interface{})
+		oldAccountIDs := make([]sdk.AccountIdentifier, len(oldAccounts))
+		for i, account := range oldAccounts {
+			oldAccountIDs[i] = sdk.NewAccountIdentifierFromAccountLocator(account.(string))
+		}
+		accountsToRemove := make([]sdk.AccountIdentifier, 0)
+		accountsToAdd := make([]sdk.AccountIdentifier, 0)
+		// Find accounts to remove
+		for _, oldAccountID := range oldAccountIDs {
+			if !slices.Contains(newAccountIDs, oldAccountID) {
+				accountsToRemove = append(accountsToRemove, oldAccountID)
+			}
+		}
+
+		// Find accounts to add
+		for _, newAccountID := range newAccountIDs {
+			if !slices.Contains(oldAccountIDs, newAccountID) {
+				accountsToAdd = append(accountsToAdd, newAccountID)
+			}
+		}
+		if len(accountsToAdd) > 0 {
+			opts := &sdk.AlterDatabaseReplicationOptions{
+				EnableReplication: &sdk.EnableReplication{
+					ToAccounts: accountsToAdd,
+				},
+			}
+			if ignoreEditionCheck := d.Get("ignore_edition_check").(bool); ignoreEditionCheck {
+				opts.EnableReplication.IgnoreEditionCheck = sdk.Bool(ignoreEditionCheck)
+			}
+			err := client.Databases.AlterReplication(ctx, id, opts)
+			if err != nil {
+				return fmt.Errorf("error enabling replication configuration on %v err = %w", d.Id(), err)
+			}
+		}
+
+		if len(accountsToRemove) > 0 {
+			opts := &sdk.AlterDatabaseReplicationOptions{
+				DisableReplication: &sdk.DisableReplication{
+					ToAccounts: accountsToRemove,
+				},
+			}
+			err := client.Databases.AlterReplication(ctx, id, opts)
+			if err != nil {
+				return fmt.Errorf("error disabling replication configuration on %v err = %w", d.Id(), err)
+			}
+		}
 	}
 
 	return ReadDatabase(d, meta)
@@ -347,18 +320,16 @@ func UpdateDatabase(d *schema.ResourceData, meta interface{}) error {
 
 func DeleteDatabase(d *schema.ResourceData, meta interface{}) error {
 	db := meta.(*sql.DB)
+	client := sdk.NewClientFromDB(db)
+	ctx := context.Background()
 	name := d.Id()
-
-	q := snowflake.NewDatabaseBuilder(name).Drop()
-	if err := snowflake.Exec(db, q); err != nil {
-		return fmt.Errorf("error deleting database %v err = %w", d.Id(), err)
+	id := sdk.NewAccountObjectIdentifier(name)
+	err := client.Databases.Drop(ctx, id, &sdk.DropDatabaseOptions{
+		IfExists: sdk.Bool(true),
+	})
+	if err != nil {
+		return err
 	}
-
 	d.SetId("")
-
 	return nil
-}
-
-func extractInterfaceFromAttribute(config interface{}, attribute string) []interface{} {
-	return config.([]interface{})[0].(map[string]interface{})[attribute].([]interface{})
 }